+++ packages/core/src/constant/constant.ts
@@ -39,6 +39,7 @@ export enum EventType {
   NODE_DBCLICK = 'node:dbclick',
   NODE_DELETE = 'node:delete',
   NODE_ADD = 'node:add',
+  NODE_DND_ADD = 'node:dnd-add',
   NODE_MOUSEDOWN = 'node:mousedown',
   NODE_DRAGSTART = 'node:dragstart',
   NODE_DRAG = 'node:drag',