+++ packages/core/src/view/behavior/DnD.ts
@@ -1,8 +1,10 @@
 import { get } from 'lodash-es';
 import LogicFlow from '../../LogicFlow';
-import { BaseNodeModel } from '../../model';
+import { EventType } from '../../constant/constant';
+
 import { TextConfig } from '../../type';
 import { snapToGrid } from '../../util/geometry';
+import { BaseNodeModel } from '../..';
 
 export type DndOptions = {
   validate: () => boolean
@@ -43,7 +45,8 @@ export default class Dnd {
     window.document.removeEventListener('mouseup', this.stopDrag);
   };
   dragEnter = (e) => {
-    if (!this.nodeConfig) return;
+    if (!this.nodeConfig || this.fakerNode) return;
+    console.log('enter');
     this.fakerNode = this.lf.createFakerNode({
       ...this.nodeConfig,
       ...this.clientToLocalPoint({ x: e.clientX, y: e.clientY }),
@@ -61,6 +64,7 @@ export default class Dnd {
   };
   onDragLeave = () => {
     if (this.fakerNode) {
+      console.log('leave');
       this.lf.removeNodeSnapLine();
       this.lf.graphModel.removeFakerNode();
       this.fakerNode = null;
@@ -70,7 +74,7 @@ export default class Dnd {
     if (!this.lf.graphModel || !e || !this.nodeConfig) {
       return;
     }
-    this.lf.addNode({
+    const currentNode = this.lf.addNode({
       ...this.nodeConfig,
       ...this.clientToLocalPoint({ x: e.clientX, y: e.clientY }),
     });
@@ -80,13 +84,17 @@ export default class Dnd {
     this.lf.removeNodeSnapLine();
     this.lf.graphModel.removeFakerNode();
     this.fakerNode = null;
+    const nodeData = currentNode.getData();
+    this.lf.eventCenter.emit(EventType.NODE_DND_ADD, { data: nodeData });
   };
 
   eventMap() {
     return {
       onMouseEnter: this.dragEnter,
+      onMouseOver: this.dragEnter, // IE11
       onMouseMove: this.onDragOver,
       onMouseLeave: this.onDragLeave,
+      // onMouseOut: this.onDragLeave, // IE11
       onMouseUp: this.onDrop,
     };
   }